+++ netbox/data_source_netbox_virtual_machines.go
@@ -120,7 +120,7 @@ func dataSourceNetboxVirtualMachine() *schema.Resource {
 							Computed: true,
 						},
 						"vcpus": {
-							Type:     schema.TypeInt,
+							Type:     schema.TypeString,
 							Computed: true,
 						},
 						"vm_id": {